Move sanitizing from cfg to invocation
