+++ libvast/src/command.cpp
@@ -14,7 +14,6 @@
 #include "vast/detail/string.hpp"
 #include "vast/detail/system.hpp"
 #include "vast/error.hpp"
-#include "vast/launch_parameter_sanitation.hpp"
 #include "vast/logger.hpp"
 #include "vast/system/application.hpp"
 #include "vast/system/start_command.hpp"
@@ -199,6 +198,42 @@ void render_parse_error(const command& cmd, const invocation& inv,
   }
 }
 
+auto generate_default_value_for_argument_type(std::string_view type_name) {
+  if (type_name.starts_with("uint") || type_name.starts_with("int")
+      || type_name.starts_with("long")) {
+    return "0";
+  } else if (type_name == "timespan") {
+    return "0s";
+  } else if (type_name.starts_with("list")) {
+    return "[]";
+  }
+  VAST_ASSERT(false && "option has type with no default value support");
+  return "";
+}
+
+void sanitize_long_form_argument(std::string& argument,
+                                 const vast::command& cmd) {
+  auto dummy_options = caf::settings{};
+  auto [state, _] = cmd.options.parse(dummy_options, {argument});
+  if (state == caf::pec::not_an_option) {
+    for (const auto& child_cmd : cmd.children) {
+      sanitize_long_form_argument(argument, *child_cmd);
+    }
+  } else if (state == caf::pec::missing_argument) {
+    auto name = argument.substr(2, argument.length() - 3);
+    auto option = cmd.options.cli_long_name_lookup(name);
+    if (!option) {
+      // something is wrong with the long name options:
+      // reveal this during the actual parsing.
+      return;
+    }
+    auto option_type = option->type_name();
+    auto options_type_default_val
+      = generate_default_value_for_argument_type(option_type.data());
+    argument.append(options_type_default_val);
+  }
+}
+
 } // namespace
 
 command::command(std::string_view name, std::string_view description,
@@ -294,11 +329,15 @@ caf::error parse_impl(invocation& result, const command& cmd,
 }
 
 caf::expected<invocation>
-parse(const command& root, command::argument_iterator first,
-      command::argument_iterator last) {
+parse(const command& root, std::vector<std::string>& arguments) {
+  for (auto &argument : arguments) {
+    if (argument.starts_with("--")) {
+      sanitize_long_form_argument(argument, root);
+    }
+  }
   invocation result;
   const command* target = nullptr;
-  if (auto err = parse_impl(result, root, first, last, &target)) {
+  if (auto err = parse_impl(result, root, arguments.begin(), arguments.end(), &target)) {
     render_parse_error(*target, result, err, std::cerr);
     return ec::silent;
   }
